Validate new todolist titles, flash messaging
Add some validation for titles of new todo lists and render error
messages on the page for invalid inputs. In addition, use
express-session and express-flash to store flash messages as session
data. This doesn't change behavior for error messages, but we can now
flash success messages through the redirect when a new todo list is
successfully created.